+++ tests/testthat/test.stamps.R
@@ -22,12 +22,12 @@ if (skip_on_cran()) {
     expect_error(mstomp.par(toy_data$data[1:200, ], window.size = 30, exc.dim = c(1, 2, 3, 4)), regexp = "exclusion dimension must be less")
 
     # small window size
-    expect_error(stamp(toy_data$data[1:200, ], window.size = 2), regexp = "Subsequence length")
-    expect_error(stamp.par(toy_data$data[1:200, ], window.size = 2), regexp = "Subsequence length")
-    expect_error(mstomp(toy_data$data[1:200, ], window.size = 2), regexp = "Subsequence length")
-    expect_error(mstomp.par(toy_data$data[1:200, ], window.size = 2), regexp = "Subsequence length")
-    expect_error(stomp(toy_data$data[1:200, ], window.size = 2), regexp = "Subsequence length")
-    expect_error(stomp.par(toy_data$data[1:200, ], window.size = 2), regexp = "Subsequence length")
+    expect_error(stamp(toy_data$data[1:200, ], window.size = 2), regexp = "Window size")
+    expect_error(stamp.par(toy_data$data[1:200, ], window.size = 2), regexp = "Window size")
+    expect_error(mstomp(toy_data$data[1:200, ], window.size = 2), regexp = "Window size")
+    expect_error(mstomp.par(toy_data$data[1:200, ], window.size = 2), regexp = "Window size")
+    expect_error(stomp(toy_data$data[1:200, ], window.size = 2), regexp = "Window size")
+    expect_error(stomp.par(toy_data$data[1:200, ], window.size = 2), regexp = "Window size")
 
     # unknown data type
     expect_error(stamp(table(rpois(100, 5)), window.size = 30), regexp = "Unknown type")
@@ -55,9 +55,9 @@ if (skip_on_cran()) {
 
   # STOMP
   stomp.test <- stomp(toy_data$data[1:200, 1], window.size = 30, verbose = 0)
-  # stomp.join.test <- stomp(toy_data$data[1:200, 1], toy_data$data[1:100, 2], window.size = 30, verbose = 0)
+  stomp.join.test <- stomp(toy_data$data[1:200, 1], toy_data$data[1:100, 2], window.size = 30, verbose = 0)
   stomp.par.test <- stomp.par(toy_data$data[1:200, 1], window.size = 30, verbose = 0)
-  # stomp.par.join.test <- stomp.par(toy_data$data[1:200, 1], toy_data$data[1:100, 2], window.size = 30, verbose = 0)
+  stomp.par.join.test <- stomp.par(toy_data$data[1:200, 1], toy_data$data[1:100, 2], window.size = 30, verbose = 0)
 
   # MSTOMP Uni
   mstomp.test1 <- mstomp(toy_data$data[1:200, 1], window.size = 30, verbose = 0)
@@ -74,8 +74,8 @@ if (skip_on_cran()) {
     test_that("Result hashes", {
       expect_known_hash(stamp.test, "1016c61c9f")
       expect_known_hash(stamp.join.test, "585be5fedc")
-      expect_known_hash(stomp.test, "9c2bf3197d")
-      # expect_known_hash(stomp.join.test, "585be5fedc")
+      expect_known_hash(stomp.test, "bd71cbd417")
+      expect_known_hash(stomp.join.test, "5521be09db")
       expect_known_hash(mstomp.test, "fa0c150b92")
       expect_known_hash(mstomp.test1, "9c2bf3197d")
       expect_known_hash(mstomp.test.must, "13cefe2517")
@@ -83,67 +83,51 @@ if (skip_on_cran()) {
     })
   }
 
-  # stamp.test -> stamp.par.test
-  # stamp.join.test -> stamp.par.join.test
-  # stamp.test -> stomp.test
-  # stamp.test -> stomp.par.test
-  # stamp.test -> mstomp.test1
-  # stamp.test -> mstomp.par.test1
-
-  # mstomp.test -> mstomp.par.test
-  # mstomp.test.must -> mstomp.par.test.must
-  # mstomp.test.exc -> mstomp.par.test.exc
-
-  # stamp.test -> stamp.par.test
+  # stamp.test and stamp.par.test
   test_that("Stamp equals to Stamp.par", {
     expect_equal(stamp.test, stamp.par.test)
   })
 
-  # stamp.join.test -> stamp.par.join.test
+  # stamp.join.test and stamp.par.join.test
   test_that("Stamp Join equals to Stamp.par Join", {
     expect_equal(stamp.join.test, stamp.par.join.test)
   })
 
-  # stamp.test -> stomp.test
+  # stamp.test and stomp.test
   test_that("Stamp equals to Stomp.par", {
     expect_equal(stamp.test, stomp.test)
   })
 
-  # test_that("Stomp equals to Stomp Join", {
-  #   expect_equal(stomp.test, stomp.join.test)
-  # })
+  test_that("Stomp Join equals to Stomp.par Join", {
+    expect_equal(stomp.join.test, stomp.par.join.test)
+  })
 
-  # test_that("Stomp Join equals to Stomp.par Join", {
-  #   expect_equal(stomp.join.test, stomp.par.join.test)
-  # })
-
-  # stamp.test -> stomp.par.test
+  # stamp.test and stomp.par.test
   test_that("Stamp equals to Stomp.par", {
     expect_equal(stamp.test, stomp.par.test)
   })
 
-  # stamp.test -> mstomp.test1
+  # stamp.test and mstomp.test1
   test_that("Stamp equals to mStomp", {
     expect_equal(stamp.test, mstomp.test1)
   })
 
-  # stamp.test -> mstomp.par.test1
+  # stamp.test and mstomp.par.test1
   test_that("Stamp equals to Stomp", {
     expect_equal(stamp.test, mstomp.par.test1)
   })
 
-
-  # mstomp.test -> mstomp.par.test
+  # mstomp.test and mstomp.par.test
   test_that("mStomp equals to mStomp.par", {
     expect_equal(mstomp.test, mstomp.par.test)
   })
 
-  # mstomp.test.must -> mstomp.par.test.must
+  # mstomp.test.must and mstomp.par.test.must
   test_that("mStomp must equals to mStomp.par must", {
     expect_equal(mstomp.test.must, mstomp.par.test.must)
   })
 
-  # mstomp.test.exc -> mstomp.par.test.exc
+  # mstomp.test.exc and mstomp.par.test.exc
   test_that("mStomp exc equals to mStomp.par exc", {
     expect_equal(mstomp.test.exc, mstomp.par.test.exc)
   })